Refactor mobymask helpers into a separate folder
