+++ packages/test/src/mobymask/helpers.ts
@@ -8,9 +8,9 @@ import {
   MOBYMASK_MESSAGE_KINDS,
   getSignedDelegationFromInvite
 } from './utils';
-import { MESSAGE_ARRIVAL_TIMEOUT, ONE_SECOND } from './constants';
-import xpaths from '../utils/elements-xpaths.json';
-import { closeDebugPanel, navigateURL, openDebugPanel, scrollElementIntoView, waitForMessage } from './driver-utils';
+import { MESSAGE_ARRIVAL_TIMEOUT, ONE_SECOND } from '../constants';
+import xpaths from './elements-xpaths.json';
+import { closeDebugPanel, navigateURL, openDebugPanel, scrollElementIntoView, waitForMessage } from '../driver-utils';
 
 export async function testPhisherReports (reportSender: WebDriver, reportReceivers: WebDriver[], phishers: string[]): Promise<void> {
   // Open debug panel on the report receivers
